+++ muna/beta/__init__.py
@@ -6,7 +6,7 @@
 from .annotations import Annotations
 from .compile import (
     compile_constant, CompileConstant, CompileDialect, CompileLibrary,
-    CompileResource, Platform
+    CompileResource, CompileTarget, Platform
 )
 from .endpoint import get_prediction_request, prediction_endpoint
 from .metadata import *