+++ muna/beta/metadata/llama.py
@@ -17,7 +17,10 @@ def _validate_llama_cpp_model(model: "llama_cpp.llama.Llama") -> "llama_cpp.llam
     except ImportError:
         raise ImportError("`llama-cpp-python` is required to create this metadata but it is not installed.")
 
-class LlamaCppInferenceMetadata(BaseModel, **ConfigDict(arbitrary_types_allowed=True, frozen=True)):
+class LlamaCppInferenceMetadata(
+    BaseModel,
+    **ConfigDict(arbitrary_types_allowed=True, frozen=True)
+):
     """
     Metadata required to lower a Llama.cpp model for LLM inference.
     """